requirements.txt
astroid==2.6.5
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
credential-shield-fdelgados==2.1.1
cryptography==3.4.7
dependency-injector==4.34.2
Faker==8.12.1
Flask==2.0.1
flask-restx==0.5.0
freezegun==1.1.0
greenlet==1.1.0
hypothesis==6.21.0
icdiff==1.9.1
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib~=0.14.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
lxml==4.6.3
packaging==21.0
parameterized==0.8.1
pika==1.2.0
pluggy==0.13.1
pprintpp==0.4.0
psutil==5.8.0
py==1.10.0
pymongo==3.12.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==6.2.4
pytest-icdiff==0.5
redis==3.5.3
requests==2.26.0
six==1.15.0
SQLAlchemy==1.4.22
tenacity==8.0.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
